send source_content for worker script
