+++ components/script/dom/worker.rs
@@ -224,17 +224,6 @@ impl WorkerMethods<crate::DomTypeHolder> for Worker {
                     devtools_sender.clone(),
                     page_info,
                 ));
-
-                let source_info = SourceInfo {
-                    url: worker_url.clone(),
-                    external: true, // Worker scripts are always external.
-                    worker_id: Some(worker_id),
-                    content: "".to_string(),
-                };
-                let _ = chan.send(ScriptToDevtoolsControlMsg::ScriptSourceLoaded(
-                    pipeline_id,
-                    source_info,
-                ));
             }
         }
